rename run eval route and remove unnecessary request parameters
